new 49th AnaTrain 19636: trying to use CabanaBoy to norm BG, still removing ghosts events, but have code to keep better track

+++ AnaTrain/Run14AuAuLeptonComby/MyEvent.h
@@ -67,7 +67,7 @@ namespace MyDileptonAnalysis
             float alpha_prime;
 
             // EMCal variables
-            short emcid;     // EMC cluster
+            int emcid;     // EMC cluster
             int emctower;  // emctower = 1000*iy + 10*iz + sector
                            // EMC sector 0-4 from bottom of arm
                            // iy tower coordinates in rphi
